fix(muya): post-process footnotes in the HTML/PDF export path
`MarkdownToHtml.renderHtml` (styled HTML / PDF / print export) ran the
marked footnote extension but never the footnote post-processing, so
exported documents left inline `[^id]` refs as literal text and footnote
definitions as raw `<div class="footnote-block">` blocks (with the
`data-identifier` already stripped by DOMPurify) — the bottom footnotes
section never appeared.

Extract `transformFootnotes` out of `renderToStaticHTML` into a shared
module and call it from the export path before sanitization, producing
the standard GFM / pandoc shape (numbered `<sup class="footnote-ref">`
markers + `<section class="footnotes">` with backrefs). Add matching
export CSS.
